\name{dsCont}
\docType{data}
\alias{dsCont}
\title{Test dataset}
\description{
  Test dataset with only continuous variables.
}
\format{Matrix containing 174 observations (rows) and 34 variables (columns).}

\author{
Gabriel Coelho Goncalves de Abreu (\email{Gabriel.Abreu@agrsci.dk}) \cr
Rodrigo Labouriau (\email{Rodrigo.Labouriau@agrsci.dk}) \cr
David Edwards (\email{David.Edwards@agrsci.dk})
}

\examples{
data(dsCont)
}

\keyword{datasets}