\name{dsDiscr}
\docType{data}
\alias{dsDiscr}
\title{Test dataset}
\description{
  Test dataset with only discrete variables.
}
\format{Data frame containing 200 observations and 20 variables.}

\author{
Gabriel Coelho Goncalves de Abreu (\email{abreu_ga@yahoo.com.br}) \cr
Rodrigo Labouriau (\email{Rodrigo.Labouriau@agrsci.dk}) \cr
David Edwards (\email{David.Edwards@agrsci.dk})
}

\examples{
data(dsDiscr)
}

\keyword{datasets}