\name{dsMixed}
\docType{data}
\alias{dsMixed}
\title{Test dataset}
\description{
  Test dataset with continuous and discrete variables.
}
\format{Data frame containing 200 observations and 15 variables.
The first 5 variables are discrete, and the last 10 continuous.}

\author{
Gabriel Coelho Goncalves de Abreu (\email{abreu_ga@yahoo.com.br}) \cr
Rodrigo Labouriau (\email{Rodrigo.Labouriau@agrsci.dk}) \cr
David Edwards (\email{David.Edwards@agrsci.dk})
}

\examples{
data(dsMixed)
}

\keyword{datasets}