\docType{package}
\name{gRapHD-package}
\alias{gRapHD-package}
\alias{gRapHD}
\title{The gRapHD package}
\description{
  The \code{gRapHD} package is designed for efficient selection of
  high-dimensional undirected graphical models. The package provides tools for
  selecting trees, forests and decomposable models minimizing information
  criteria such as AIC or BIC, and for displaying the independence graphs of the
  models. It has also some useful tools for analysing graphical structures. It
  supports the use of discrete, continuous, or both types of variables.
}

\author{
Gabriel Coelho Goncalves de Abreu (\email{Gabriel.Abreu@agrsci.dk}) \cr
Rodrigo Labouriau (\email{Rodrigo.Labouriau@agrsci.dk}) \cr
David Edwards (\email{David.Edwards@agrsci.dk})
}
