\name{print.gRapHD}
\alias{print.gRapHD}
\title{Prints an object of class "gRapHD".}
\description{
  \code{\link{Methods}} for class \code{gRapHD}.
}

\usage{
  \method{print}{gRapHD}(x,...)
}

\arguments{
  \item{x}{an object of class \code{gRapHD}.}
  \item{\dots}{further arguments passed to or from other methods.}
}

\details{
  So far it only presents a summary of the object.
}

\author{
Gabriel Coelho Goncalves de Abreu (\email{Gabriel.Abreu@agrsci.dk}) \cr
Rodrigo Labouriau (\email{Rodrigo.Labouriau@agrsci.dk}) \cr
David Edwards (\email{David.Edwards@agrsci.dk})
}

\examples{
  m <- as.gRapHD(NULL)
  m

}
\keyword{graphs}
